requirements.txt
torch>=2.1.0
torchvision>=0.16.0
nibabel>=5.2.0
SimpleITK>=2.3.0
scikit-image>=0.22.0
scipy>=1.12.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
PyYAML>=6.0
tqdm>=4.66.0
tensorboard>=2.15.0
